requirements/prod.txt
Django==4.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.11
django-silk==5.3.2
djangorestframework==3.15.2
drf-yasg==1.21.8
pillow==11.0.0
psycopg2-binary==2.9.10
